requirements.txt
faster_whisper
ollama
pydub
soundfile
numpy
python-dotenv
torch==2.5.1 
torchvision==0.20.1 
torchaudio==2.5.1
TTS